requirements.txt
langchain
langchain-community
langchain-groq
openai
python-dotenv
pandas
matplotlib
seaborn
plotly
